refactor: SimpleButton.vueをatoms/へ
