+++ src/pages/index.vue
@@ -13,7 +13,7 @@ meta:
 import ArticleCardGrid from "@/components/ArticleCardGrid.vue";
 import PageSet from "@/modules/PageSet";
 import SzppIcon from "@/components/atoms/SzppIcon.vue";
-import SimpleButton from "@/components/SimpleButton.vue";
+import SimpleButton from "@/components/atoms/SimpleButton.vue";
 import FlexBox from "@/components/FlexBox.vue";
 </script>
